Adding JSValueType as a separate type
The C++ side has two separate enums - ValueType and JSValueType.
ValueType has all the variants as JSValueType except Unknown. There may
be a smarter way to merge these two types with an assertion that the
JSValueType is not Unknown ? but adding it as a new type for now to
avoid "unsafe" casts on the C++ implementation side.